Rare Write support for statically allocated memory
The options currently available for protecting memory are limited to:
* const: memory is write protected from early boot stages.
* __ro_after_init: memory is writable during the init phase.

Which means that if some data is modified very seldomly (or even just
once), but after init, it is not possible to give it any additional level
of protection.

Reasons for additional protection:
* accidental overwrites due to bugs / malfunctions.
* malicious attacks, exploiting some vulnerability.

The former type of protection is beneficial for any data.
The latter type is more relevant for data that might be targeted by an
attack to the system integrity.

The most common type of attack aims at flipping some kernel state variable,
often statically allocated. The purpose of the attack is to alter the
execution flow, for example to bypass existing safety measures.

These types of attack rely on having read/write capability to kernel
memory (write is typically limited to unprotected kernel data).
Such attacks will usually rely on identifying a vulnerability in some
kernel API, then exploit it, to read/alter arbitrary locations.

This patch is intended to prevent such direct attacks.

However, it doesn't prevent more sophisticated attacks, based, for
example, on walking "manually" the page table and altering attributes of
victim pages.

(Note: what was just described is an intrinsic vulnerability that
affects also __ro_after_init.
const data is less affected because the compiler can perform additional
optimizations and replace references to the constant with an immediate,
where possible.)

Nor, does this patch prevent Control Flow based attacks.
For those, a specific compiler plugin is needed.

However, even with these limitations, the patch still provides higher
protection than the current state of affairs, where such variables are
directly writable, all the time.

But the patch must not introduce new weaknesses, nor compromise overall
performance.

The mechanism used for performing the alteration (rare write) is to:
* identify the page - or pages - targeted
* sequentially, remap them to a random location, with write permission
* alter the interested locations
* destroy the writable mapping

To avoid introducing weaknesses:
* the rewritable memory is contained in a subsection with boundaries
* prior to any change, it is confirmed to belong to the rewritable section
* the altering functionality is inlined, to minimize code reuse attacks

To avoid compromising performance, the patch does not introduce any
locking.

As long as multiple writers do not write to the same memory
locations, there is no need to introduce serialization.
If two or more cores/threads were to map the same page to different
addresses, writing to different locations within the same page, the cache
coherency is still performed at physical level, making sure to
propagagate the changes.

The individual execution of a rare-write operation has, of course,
longer duration than a simple assignement to unprotected memory, however
it doesn't require stalling any other core/thread, therefore leaving any
penalty only on the core/thread performing the rare-write.

The special modifier added is similar to __ro_after_init, with a similar
name and behavior: __rare_write_after_init.

For the sake of having one single implementation of the functionality,
working with both linear and virtually linear memory, the core
rare_write function supports selection of memory type, which gets
optimized away at compile time once a specific type is chosen.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>
CC: Carlos Chinea Perez <carlos.chinea.perez@huawei.com>
CC: Remi Denis Courmont <remi.denis.courmont@huawei.com>
CC: Ashish Raj <araj@elisanet.fi>