Rare Write support for statically allocated memory
The options currently available for protecting statically allocated memory
are limited to:
* const: memory is write protected from early boot stages.
* __ro_after_init: memory is writable during the init phase.

Data which is modified very seldomly (or even just once), but after init,
cannot benefit from any additional level of protection.

Reasons for wanting additional protection on the data:
* accidental overwrites due to bugs / malfunctions.
* malicious attacks, typically from userspace, exploiting some vulnerability.

The former type of protection is beneficial for any data.
The latter type is more relevant for data that might be targeted by an
attack to the system integrity.

The most common type of attack aims at flipping some kernel state variable,
often statically allocated. The purpose of the attack is to alter the
execution flow, for example to bypass existing safety measures.

These types of attack rely on having read/write capability to kernel
memory (write is typically limited to unprotected kernel data).
Such attacks will usually exploit an unpatched vulnerability in some
kernel API, then exploit it, to read/alter arbitrary locations.

This patch is intended to make such direct attacks harder to perform.

Harder but not impossible, because if the attacker can read/write kernel
memory, there is still the possibility to undo whatever protection the
kernel might have setup, for example using jop/rop or a
page-reprogramming algorythm from user space.

This sort of intrinsic vulnerability affects also __ro_after_init,
though.

const data is less affected in those cases where the compiler can perform
additional optimizations and replace references to the constant with an
immediate value. In every other case, it's equally affected.

Nor, does this patch prevent Control Flow based attacks.
For those, other means are required, such as specific compiler plugins and
additional support from the hardware, like the features surfacing on the
most recent ARM and Intel architectures.

However, even with these limitations, the patch still provides higher
protection than the current state of affairs, where such variables are
directly writable, all the time.
And the attacker is forced to execute far more difficult tasks, which
require specific actions that could be detected more easily, like a page
walk performed from userspace.

But the patch must not introduce new weaknesses, nor compromise overall
performance.

The mechanism used for performing the alteration (rare write) is to:
* identify the page - or pages - targeted
* one by one, remap them to a random location, with write permission
* alter the interested locations
* destroy the writable mapping

To avoid introducing weaknesses:
* the rewritable memory is contained in a subsection with boundaries
* prior to any change, it is confirmed to belong to the rewritable section
* the altering functionality is inlined, to minimize code reuse attacks
  by allowing the compiler to generate less-generic code.

The reason for performing the remapping is that static variables are the
easiest to locate in memory and, therefore, unprotecting the active
mapping would be risky, at best.
By directly unprotecting the primary mapping, an attacker might try to busy
loop on another core/thread, while a rare write is ongoing, to take
advantage of the time window when the data is writable, at its primary
address. Furthermore, the attacker could alter anything that is contained
in the same page of the data being written to.
However, by remapping the page to a separate address, the protections on
its initial mapping is preserved and the attacker has the much harder task
to figure out where the page is being remapped, before being able to
take advantage of it.

To prevent an unlikely - but possible - attack from irq context, local
irqs are disabled for the duration of a remapping.

Other types of attack, which might try to reduce the randomness of the
remapping are still possible and should be addressed separately as any
attack that tries to reduce the entropy of the source of randomness.

The patch does not affect overall performance, since it doesn't introduce
any global locking, to keep other cores at bay.
It has limited impact on responsiveness to interrupts, which can be
quantified as the time taken for establishing the new mapping, copying
the data (max 1 page) and tearing down the temporary mapping.

From the perspective of ensuring integrity, in a scenario with
multiple writers, as long as they do not try write to the same memory
locations, there is no need to introduce serialization.

If two or more cores/threads were to map the same page to different
addresses, writing to different locations within the same page, the cache
coherency mechanism, which is performed at physical level, would ensure to
propagate the changes.
Locking is needed only in the typical case where multiple writers want
to alter the very same memory, but that is not different from any normal
memory assignment.

The individual execution of a rare-write operation has, of course,
longer duration than a simple assignement to unprotected memory, however
it doesn't require stalling any other core/thread, therefore leaving any
penalty only on the core/thread performing the rare-write.

Writes are performed through __write_once_size(), to ensure that atomicy
is preserved, where it might matter (ex: update of a pointer).

The special modifier added is similar in both name and behavior, to
__ro_after_init: __rare_write_after_init.

However, there is one major difference: in this case the interval that
is reserved for static rare writes MUST be page aligned, so that other
types of data do not share any page, to avoid mixing data types with
different access patterns, which would either leave unprotected some
data that is suppoed to be read-only, or it would write protect some
other data which is supposed to be always writable.

For the sake of having one single implementation of the functionality,
working with both linear and virtually linear memory, the core
rare_write function supports selection of memory type, which gets
optimized away at compile time once a specific type is chosen.
This is needed to invoke the correct function for obtaining the page
structure, without adding extra runtime tests. Furthermore it reduces
the chances that an attacker would use it with the unintended type of
addresses.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>
CC: Ashish Raj <araj@elisanet.fi>
CC: Carlos Chinea Perez <carlos.chinea.perez@huawei.com>
CC: Remi Denis Courmont <remi.denis.courmont@huawei.com>