Protection for dynamically allocated memory
Currently, memory allocated dynamically is never safeguarded against
unwanted alterations, either accidental or malicious.

This is different from what is available for statically allocated
variables, which can be write protected:
* either against any further modification (__read_only_after_init)
* or against direct writes, leaving the possibility to be altered
  indirectly, albeit in a more costly way (__rare_write_after_init)

The problem with protecting dynamically allocated memory is that the HW
protection works only at a page-level granularity.

For statically allocated variables, the packing of variables to protect
happens by indirectly telling the linker where to place the variables that
need protection. It is done by using either of the __read_only_after_init
and __rare_write_after_init modifiers.

However, this sort of discrimination is not available for variables
allocated dynamically.
When invoking kmalloc and friends, there is almost no control whatsoever on
where the memory will be coming from, since they do their best to slice,
dice and dish out memory in a way that keeps fragmentation low, thus
packing together miscellaneous requests.

If the memory area being write-protected happens to contain also some
other data which will need to be altered again, this will cause errors.

One possible option would have been to have specific memory zones.
However it seems very hard, if not impossible, to obtain support for two
new memory zones. It would be also somewhat problematic to define
upfront the size of such zones.

The next best thing seems to be using vmalloc-backed allocations.
Vmalloc memory has the nice property of being virtually contiguous,
meaning that at any point in time, as long as there are enough free
pages, any request will be satisfied.

This patch introduces pools of protected/protectable memory, implemented
as lists of vmap_areas.

Only data which is intended to be write protected will be allocated
through these pools, ensuring phisical separation from other which is
incompatible with the write-protection.

There are, however, two major drawbacks in using plain vmalloc:

1) vmalloc uses one extra layer of mapping, meaning that each page has
its own TLB entry, instead of the continuous mapping that is possible with
the linear memory. This can cause TLB trashing, when in presence of a large
number of small allocations, which is the typical case when converting
existing code that relies on kmalloc.

2) vmalloc allocates at least one new page every time it is invoked - in
practice it is a wrapper for get_free_pages.
Vmalloc stitches the pages together, in the virtual address space.
If an allocation request is only for a handful of bytes, vmalloc will still
return a whole page, with a large amount of wasted space. It might not seem
a problem on a big iron system, where there is anyway a very large amount
of memory available, however on an embedded/IoT class device, memory can
be a very scarce resource.

Taking as example the list of measurements kept by IMA, each element
being allocated is composed by 4 pointers, which on a 64 bit system amount
to 4 x 8 = 32 bytes. Taking the smallest PAGE_SIZE available, of 4KB,
it means that the reserved:used ratio is 128:1.
Assuming a filesystem with 1000 entries to be measured, it means going
from less than 8 pages (32KB) and 8 TLB entries, to 1000 pages (almost 4MB)
and 1000 TLB entries.

The situation would become even worse when trying to protect the SELinux
policyDB: on a mobile phone it can use few MB of linear memory.
With a kmalloced policyDB of, say, 5MB, subdivided in allocation units
that average to 32Bytes or less, it would translate to more than 160.000
allocations, using 1 TLB each, for a grand total of 640MB, using 4KB
pages.

This clearly is not feasible and must be addressed, if vmalloc is to be
used as memory provider.

The solution is to have a linear allocator layered on top of vmalloc:
pmalloc. Pmalloc reuses the current vmap_area, as long as the residual
space it contains is large enough to accomodate the current allocation
request.

If the space is insufficient, it will allocate the larger between the
default allocation size and the memory requested, rounded up to a
multiple of PAGE_SIZE.

Such type of allocator is even preferrable to the buddy allocator used
by kmalloc, under certain circumstances.
In the IMA case, for example, the memory allocated will never be
released, leading to a progressive fragmentation of the SLAB, since the
buddy allocator will never be able to merge back those memory chunks
and, not being aware of it, it will pick them from whatever page is
available at the moment, effectively

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>