Protectable memory allocator
Currently, memory allocated dynamically is never safeguarded against
unwanted alterations, either accidental or malicious.

This is different from what is available for statically allocated
variables, which can be write protected:
* either against any further modification (__ro_after_init)
* or against direct writes, leaving the possibility to be altered
  indirectly, albeit in a more costly way (__wr_after_init)

The problem with protecting dynamically allocated memory is that the HW
protection works only at a page-level granularity.

For statically allocated variables, the packing of protected variables
happens by indirectly telling the linker where to place the variables that
need protection. It is done by using either of the __ro_after_init and
__wr_after_init modifiers.

However, this sort of discrimination is not available for variables
allocated dynamically.
When invoking kmalloc and friends, there is almost no control about the
origin of the memory they provide. It is possible to request from a certain
memory zone, but that has little to do with specifying the memory page.
Such allocators do their best to slice, dice and dish out memory in a way
that keeps fragmentation low, thus packing together miscellaneous requests,
caring little about how the memory will be used and mostly about the size
requested.

If the memory pages containing the data that should be write-protected
happen to contain also some other data that gets occasionally altered, the
protection is impossible, because it would cause write errors, whenever
such legitimate alterations take place.

One possible option would have been to have specific memory zones.
However it seems very hard, if not impossible, to obtain inrastructure
support for two new memory zones. It would be also somewhat problematic to
define upfront the size of such zones.

Using vmalloc-backed allocations seems to be the next best thing.
Vmalloc memory has the nice property of being virtually contiguous,
meaning that at any point in time, as long as there are enough free
pages, any request will be satisfied.

This patch introduces pools of protected/protectable memory, implemented
as lists of vmap_areas.

These pools are exclusively for data requiring write protection.
They prevent the random interleaving, within the same pages and vmap_areas,
of both write protected and regular data.

There are, however, two major drawbacks in using plain vmalloc:

1) vmalloc allocates at least one new page every time it is invoked - in
practice it is a wrapper for get_free_pages, adding contiguity.
Vmalloc stitches the pages together, making them look contiguous.
Even if an allocation request demands only a handful of bytes, vmalloc will
still return a whole page, which will not be used for any other purpose,
even if it is mostly unused, wasting a realtively large amount of space.
It might not seem a problem on a big iron system, where there is anyway a
very large amount of memory available, however on an embedded/IoT class
device, memory can be a very scarce resource. And even in the case of a
large system, it would be potentially very wasteful.

2) vmalloc pages are typically not physically contiguous, meaning that each
page has its own TLB entry, instead of the contiguous mapping that is
possible with sections of linear memory and requires far less TLB
entries.
This can cause TLB trashing, when in presence of a large number of small
allocations, which is the typical case, when converting existing code that
relies on kmalloc as memory backend.

Taking as example the list of measurements kept by IMA, each element
being allocated is composed by 4 pointers, which on a 64 bit system amount
to 4 x 8 = 32 bytes. Even considering the smallest PAGE_SIZE available,
4KB, it means that, when transitioning from kmalloc to vmalloc, the
reserved:used ratio is 128:1.
Assuming a filesystem with 1000 entries to be measured, it means going
from less than 8 pages (32KB) and 8 TLB entries (with kmalloc),
to 1000 pages (almost 4MB) and 1000 TLB entries (with plain vmalloc).

The situation would become even worse, when trying to protect the SELinux
policyDB: on a mobile phone it can use few MB of linear memory.
With a kmalloced policyDB of, say, 5MB, subdivided in allocation units
that average to 32Bytes or less, it would translate to more than 160.000
allocations, using 1 TLB each, for a grand total of 640MB, using 4KB
pages.

This clearly is not feasible and must be addressed, if vmalloc is to be
used as memory provider.

The solution is to have a linear allocator layered on top of vmalloc:
pmalloc. When pmalloc otains a vmap_area from vmalloc, it keeps using it
as source of free memory, as long as the residual free space in the
vmap_area is large enough to accommodate the current allocation request.

If the space in the currently available vmap_area is insufficient, pmalloc
will allocate another, with a size that is the larger between the default
allocation size and the memory requested, rounded up to a multiple of
PAGE_SIZE.

In case of large allocations, it will not perform worse than the plain
vmalloc, which it still relies on.

Such type of allocator is even preferrable to the buddy allocator used
by kmalloc, under certain circumstances.
In the IMA case, for example, the memory allocated will never be
released, leading to a progressive fragmentation of the SLAB, since the
buddy allocator will never be able to merge back those memory chunks
and, not being aware of it, it will pick them from whatever page is
available at the moment, effectively

The pools support two main types of protection: read_only and write_rare.

Read-only is for data that has a write-once life cycle, at most it can
be destroyed when not needed anymore, like the SElinux PolicyDB.

Write Rare, instead, is more suitable for data which occasionally must
be modified, like the tail of the IMA measurement list.

Pools can also be configured to perform auto-locking and start already
write protected (obviously it makes sense only for write rare).
Auto locking can be convenient, when only the latest allocation needs to be
directly writable.

Starting already locked, instead provides immediately protection for
data that might be otherwise under attack, during its initialization phase.

It must also be possible to destroy pools, when they are not needed
anymore, for example when unloading a module that created the pool, or
when re-initializing the data structure(s) it contains, like when the
SELinux PolicyDB is reloaded.

About the write rare implementation: during a discussion with Matthew
Wilcox and Dave Hansen, at LFSMM 2018, it was proposed and agreed that
rare write would be implemented by writing directly to the linear mapping,
because I was under the impression that this was not protected.
However, when trying to implement this solution, it turned out that, at
least on x86 and arm64, it is indeed already protected.
OTOH, the implementation of write rare for linear mapping is already coping
with this situation, so it seemed preferrable to keep only one
implementation for write rare and reuse it.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>