prmem: write rare for static allocation
Implementation of write rare for statically allocated data, located in a
specific memory section through the use of the __write_rare label.

The basic functions are:
- wr_memset(): the write rare counterpart of memset()
- wr_memcpy(): the write rare counterpart of memcpy()
- wr_assign(): the write rare counterpart of the assignment ('=') operator

The implementation is based on code from Andy Lutomirski and Nadav Amit
for patching the text on x86 [here goes reference to commits, once merged]

The modification of write protected data is done through a temporary,
writable, local mapping, accessible only through a separate mm struct, to
the core that is performing the write operation.

Local interrupts are disabled, while the alternate mapping is active.
When the destination crosses page boundaries, only one page is re-mapped at
any point in time.

When transitioning to the next page, after the old remap is torn down and
before the new one is created, interrupts are enabled, to give a chance to
pending ones of being serverd.

They are subsequently disabled again and the next re-mapping resumes the
cycle, but this guarrantees a deterministic upper bound to the maximum
latency.

Signed-off-by: Igor Stoppa <igor.stoppa@huawei.com>
CC: Andy Lutomirski <luto@amacapital.net>
CC: Nadav Amit <nadav.amit@gmail.com>
CC: Matthew Wilcox <willy@infradead.org>
CC: Peter Zijlstra <peterz@infradead.org>
CC: Kees Cook <keescook@chromium.org>
CC: kernel-hardening@lists.openwall.com
CC: linux-kernel@vger.kernel.org